fix: start standardizing trackgroup responses
